update links and displayNames
